BLSS: turn the activation table on, and re-tune the demo until a person can watch it
Three things, and the third is the one that changes what we claim.

**The activation table ships.** `TYRA_BLSS_ACT_TABLE` (engine) and `actTable`
(src/blss.cpp) are one number in two files; both move to 512 here, in the same
commit, because a one-sided flip is silent twin divergence - the exact failure
docs/blss-reconstruction.md exists to prevent. Parity checked with
`--blss-eval <projectDir> -i <net>`: the headline table is unchanged, two
per-shot columns move by 0.001 dB, which is what a max deviation of 1.5e-05 from
tanh predicts. On hardware it takes `net` from **1.93 to 0.79 ms** - PCSX2 had
priced it at 2.11 and the hardware bound was stated as "at most ~1.5"; it is
1.14. BLSS' EE bill is now **5.02 ms**.

**examples/upscaler-lab is re-tuned against the console.** It measured 3.37x
faster with BLSS on and ran at **1.9 FPS**, because it had been tuned against
PCSX2's FPS counter and PCSX2 under-reports GS fill by 76x. 12 banks x 256
billboards become **6 x 32** - same `size`, same material, same cottages,
props, spiders, campfires, rain and 20 s tour. Measured on a real PS2, paired
`FTRAW` frames from the parked-camera region, two runs per arm, all four
cross-pairings:

    BLSS off   52.86 ms   median 53.01   p95 54.44   18.9 FPS
    BLSS on    32.98 ms   median 32.93   p95 33.62   30.3 FPS
    d = +19.88 ms, 95 % CI [+19.81, +19.95], sd 1.12, n = 1024, 1.60x

The two fixtures differ only in particle count, so together they fit the fill
model: **BLSS keeps 25.9 % of the scene's fill** (`blssScale 0` is `Scale::X2Y2`
- half in *each* axis, a quarter of the pixels), and the scene's non-haze floor
is 21.2 ms off / 24.8 ms on. That floor is why the demo lands at 19 -> 30 FPS
and not at the 20 -> 45 that was wanted: thinning the haze further only shrinks
the win, because the haze is the only thing BLSS is paid to remove.

**The docs led with a regression that is no longer the result.** They said
"proof of concept, do not enable this yet" and a provisional +9.83 ms. Rewritten
around what is measured: the feature is a large win above a break-even overdraw
and a loss below it, the break-even is **~13 full-screen alpha-blended
coverages** (the old ~22 was computed against a 2x resolution factor and the
pre-table EE bill), the regime is heavy alpha-blended overdraw, and both sides
of the line are measured on hardware. The `drain` discriminator error is kept on
the page, because it is more instructive than the conclusion it produced.

Also: the net is retrained and recommitted. The re-tune did **not** change the
training corpus - `blssscene` never sees an emitter, so it stayed byte-identical
at 156 frames / 34 944 tiles and the oracle ceiling is still +1.730 dB - but the
activation table changes the forward pass, and the refit came out slightly
better: **+0.85 dB against +0.78**, 49 % of the ceiling against 45 %. The
example's "textured primitives vanish with BLSS on" defect no longer reproduces
and is marked fixed; the period-2 bob does, and that example still keeps
`blssJitter: true` deliberately as the reference for it, so set it false before
demonstrating the scene to anyone.

+++ examples/upscaler-lab/src/terrain_game.cpp
@@ -1489,12 +1489,11 @@ constexpr float kTicksPerMs = 294912.0F;
 constexpr u32 kBudget20ms = 5898240U;  // 20 ms of COP0 Count = the PAL budget
 
 u32 work[kWindow], drain[kWindow];
-u32 sBeg = 0, sEnd = 0, sCmp = 0, sCmpEe = 0;
-// The split of the two big EE terms - proxy feed / reprojection / features /
-// MLP / packet build. The first A/B read the composite's EE half off ONE
-// counter and got "~3.9 ms of scene submission" by SUBTRACTION, which is not
-// a measurement; these make both attributable.
-u32 sPrx = 0, sRep = 0, sFea = 0, sNet = 0, sPkt = 0;
+u64 sBeg = 0, sEnd = 0, sCmp = 0, sCmpEe = 0;
+// The split of the two big EE terms. The first hardware A/B read the
+// composite's EE half off ONE counter and got "~3.9 ms of scene submission"
+// by SUBTRACTION, which is not a measurement; these make both attributable.
+u64 sPrx = 0, sRep = 0, sFea = 0, sNet = 0, sPkt = 0;
 int n = 0;
 u32 frame = 0;  // frames since boot - the alignment key between runs A and B
 u32 raw[kRaw];
@@ -1508,7 +1507,12 @@ u32 pBeg = 0, pEnd = 0, pCmp = 0, pCmpEe = 0;
 u32 pPrx = 0, pRep = 0, pFea = 0, pNet = 0, pPkt = 0;
 bool pValid = false;
 
-inline float ms(u32 ticks, int count) {
+// u64, because a u32 SUM OVERFLOWS. 50 frames x 300 ms is 4.4e9 ticks against
+// a 4.29e9 ceiling, so on a scene slow enough to be worth profiling the mean
+// wrapped and printed BELOW the median - which is how a 500 ms frame first
+// reported itself as 37 ms. Per-frame values (median, p95) were always fine;
+// only the accumulators were wrong.
+inline float ms(u64 ticks, int count) {
   return (float)ticks / (kTicksPerMs * (float)count);
 }
 
@@ -1575,7 +1579,7 @@ void tick(const Vec4& camPos, const Vec4& camAt) {
   n = 0;
 
   u32 sorted[kWindow];
-  u32 sumW = 0, sumD = 0;
+  u64 sumW = 0, sumD = 0;
   int over = 0;
   for (int k = 0; k < kWindow; k++) {
     sorted[k] = work[k];
@@ -1603,10 +1607,10 @@ void tick(const Vec4& camPos, const Vec4& camAt) {
            (double)ms(sCmp, kWindow), (double)ms(sCmpEe, kWindow),
            (double)ms(sCmp - sCmpEe, kWindow), over, (double)cam);
   TYRA_LOG(line);
-  // The attribution line. proxy is charged inside StaPipCore (scene
-  // submission, NOT the composite); the other four are the composite's EE
-  // half split at its four phases, so reproj+feat+net+pkt should reconstruct
-  // comp's EE figure above to within the counters' own overhead.
+  // The attribution line. `proxy` is charged inside StaPipCore (scene
+  // SUBMISSION, not the composite); the other four split the composite's EE
+  // half at its four phases, so reproj+feat+net+pkt reconstructs comp's EE
+  // figure above to within the counters' own overhead.
   snprintf(line, sizeof(line),
            "FTSPLIT f=%lu proxy=%.2f reproj=%.2f feat=%.2f net=%.2f pkt=%.2f",
            (unsigned long)(frame - kWindow), (double)ms(sPrx, kWindow),
